fix: mining MVP runtime hardening and expanded test coverage
- Validate unrevealed sectors cannot carry a mine in save repository
- Ensure simulation arithmetic stays as double after clamping
- Add missing switch breaks in mining action sheet
- Use Paint.from(accent) to avoid mutating shared paint
- Guard camera transitions until game has loaded
- Include OperationLightComponent in sector structure query
- Simplify pan delta with unary negation
- Expand mining integration and unit tests

+++ lib/mining/mining_simulation.dart
@@ -58,9 +58,12 @@ class MiningSimulation {
       if (!progress.revealed || mine == null) continue;
 
       final capacity = content.capacityFor(definition.id, mine.level);
-      final remaining = (capacity - mine.storedAmount).clamp(0.0, capacity);
+      final remaining = (capacity - mine.storedAmount)
+          .clamp(0.0, capacity)
+          .toDouble();
       final amount = (content.rateFor(definition.id, mine.level) * seconds)
-          .clamp(0.0, remaining);
+          .clamp(0.0, remaining)
+          .toDouble();
       final stored = mine.storedAmount + amount;
       sectors[definition.id] = progress.copyWith(
         mine: mine.copyWith(storedAmount: stored),
